+++ scripts/build-pandas.sh
@@ -5,10 +5,14 @@ set -e
 . "${SCRIPTS_DIR}/environment.sh"
 
 cd "${SOURCES_DIR}/${3}"
-git clean -xdf
-git reset --hard
-git checkout "${2}"
-python3 setup.py bdist
+
+if ! [ -d "build" ]; then
+  git clean -xdf
+  git reset --hard
+  git checkout "${2}"
+  python3 setup.py bdist
+fi
+
 find ./build/lib*/"${1}" -name '*-darwin.so' -exec sh -c 'mv "$0" "${0/darwin.so/iphoneos.dylib}"' {} \;
 make-frameworks.sh --bundle-identifier "org.pydata" --bundle-name "${1}" --bundle-version "${2}" --input-dir ./build/lib*/"${1}" --output-dir "${FRAMEWORKS_DIR}"
 cp -r build/lib*/* "${SITE_PACKAGES_DIR}/${1}"
