+++ scripts/build-pywt.sh
@@ -8,10 +8,14 @@ export NPY_BLAS_ORDER=""
 export NPY_LAPACK_ORDER=""
 
 cd "${SOURCES_DIR}/${3}"
-git clean -xdf
-git reset --hard
-git checkout "${2}"
-python3 setup.py bdist --force
+
+if ! [ -d "build" ]; then
+  git clean -xdf
+  git reset --hard
+  git checkout "${2}"
+  python3 setup.py bdist --force
+fi
+
 find ./build/lib*/"${1}" -name '*-darwin.so' -exec sh -c 'mv "$0" "${0/darwin.so/iphoneos.dylib}"' {} \;
 make-frameworks.sh --bundle-identifier "com.github" --bundle-name "${1}" --bundle-version "${2}" --input-dir ./build/lib*/"${1}" --output-dir "${FRAMEWORKS_DIR}"
 cp -r build/lib*/* "${SITE_PACKAGES_DIR}/${1}"
