+++ scripts/build-statsmodels.sh
@@ -7,19 +7,22 @@ set -e
 export LDFLAGS="-L'${FRAMEWORKS_DIR}'"
 
 cd "${SOURCES_DIR}/${3}"
-git clean -xdf
-git reset --hard
-git checkout "${2}"
-python3 setup.py bdist
 
-src_dir="build/temp.iphoneos-arm64-3.8"
-dst_dir="build/lib.iphoneos-arm64-3.8"
+if ! [ -d "build" ]; then
+  git clean -xdf
+  git reset --hard
+  git checkout "${2}"
+  python3 setup.py bdist
 
-find "${src_dir}" -type f -name "*.o" | while read -r file; do
+  src_dir="build/temp.iphoneos-arm64-3.8"
+  dst_dir="build/lib.iphoneos-arm64-3.8"
+
+  find "${src_dir}" -type f -name "*.o" | while read -r file; do
     rel_path="${file#"${src_dir}"/statsmodels/}"
     out_path="${dst_dir}/statsmodels/${rel_path/.o/.cpython-38-darwin.so}"
     "${CC}" "${LDFLAGS}" -undefined dynamic_lookup -bundle "${file}" -o "${out_path}"
-done
+  done
+fi
 
 find ./build/lib*/"${1}" -name '*-darwin.so' -exec sh -c 'mv "$0" "${0/darwin.so/iphoneos.dylib}"' {} \;
 make-frameworks.sh --bundle-identifier "org" --bundle-name "${1}" --bundle-version "${2}" --input-dir ./build/lib*/"${1}" --output-dir "${FRAMEWORKS_DIR}"
